Store unresolved claims
In a later commit those will be processed when a new token network is
registered.
