+++ raiden/claim.py
@@ -1,102 +1,43 @@
 import json
-from dataclasses import dataclass
 from pathlib import Path
 
-from eth_abi import encode_single
-from eth_utils import keccak, to_canonical_address, to_hex
-from web3 import Web3
+from eth_utils import to_canonical_address
 
 from raiden.settings import MediationFeeConfig
 from raiden.storage.serialization import DictSerializer
+from raiden.transfer import views
 from raiden.transfer.architecture import StateChange
 from raiden.transfer.identifiers import CanonicalIdentifier
 from raiden.transfer.mediated_transfer.mediation_fee import FeeScheduleState
 from raiden.transfer.state import (
+    ChainState,
+    Claim,
     NettingChannelEndState,
     NettingChannelState,
     SuccessfulTransactionState,
-    TokenNetworkGraphState,
-    TokenNetworkState,
     TransactionChannelDeposit,
 )
 from raiden.transfer.state_change import (
     ContractReceiveChannelDeposit,
     ContractReceiveChannelNew,
-    ContractReceiveNewTokenNetwork,
     ContractReceiveRouteNew,
 )
-from raiden.utils.formatting import to_checksum_address, to_hex_address
-from raiden.utils.signer import Signer
 from raiden.utils.typing import (
     Address,
-    Any,
     Balance,
     BlockHash,
     BlockNumber,
     BlockTimeout,
-    Dict,
     List,
-    Optional,
-    TokenAddress,
     TokenNetworkAddress,
     TokenNetworkRegistryAddress,
     TransactionHash,
 )
-from raiden_contracts.utils.type_aliases import ChainID, ChannelID, Signature, TokenAmount
+from raiden_contracts.utils.type_aliases import ChainID, TokenAmount
 
 CLAIM_FILE_PATH = Path("./claims.json")
 DEFAULT_SETTLE_TIMEOUT = BlockTimeout(100)
 DEFAULT_REVEAL_TIMEOUT = BlockTimeout(50)
-
-
-@dataclass
-class Claim:
-    chain_id: ChainID
-    token_network_address: TokenNetworkAddress
-    owner: Address
-    partner: Address
-    total_amount: TokenAmount
-    signature: Optional[Signature] = None
-
-    def pack(self) -> bytes:
-        return (
-            Web3.toBytes(hexstr=to_hex_address(self.token_network_address))
-            + encode_single("uint256", self.chain_id)
-            + Web3.toBytes(hexstr=to_hex_address(self.owner))
-            + Web3.toBytes(hexstr=to_hex_address(self.partner))
-            + encode_single("uint256", self.total_amount)
-        )
-
-    def sign(self, signer: Signer) -> None:
-        self.signature = signer.sign(data=self.pack())
-
-    def serialize(self) -> Dict[str, Any]:
-        assert self.signature is not None, "Claim not signed yet"
-        return dict(
-            chain_id=self.chain_id,
-            token_network_address=to_checksum_address(self.token_network_address),
-            owner=to_checksum_address(self.owner),
-            partner=to_checksum_address(self.partner),
-            total_amount=self.total_amount,
-            signature=to_hex(self.signature),
-        )
-
-    def channel_id(self) -> ChannelID:
-        if self.owner < self.partner:
-            hashed_id = keccak(
-                encode_single("uint256", self.chain_id)
-                + self.token_network_address
-                + self.owner
-                + self.partner
-            )
-        else:
-            hashed_id = keccak(
-                encode_single("uint256", self.chain_id)
-                + self.token_network_address
-                + self.partner
-                + self.owner
-            )
-        return ChannelID(int.from_bytes(bytes=hashed_id, byteorder="big"))
 
 
 _DUMMY_ADDRESS = to_canonical_address("0x" + "0" * 40)
@@ -123,32 +64,21 @@ def parse_claims_file() -> List[Claim]:
     ]
 
 
-def claims_to_blockchain_events(
+def get_state_changes_for_claims(
+    chain_state: ChainState,
     claims: List[Claim],
     node_address: Address,
     token_network_registry_address: TokenNetworkRegistryAddress,
-    token_network_address: TokenNetworkAddress,
-    token_address: TokenAddress,
 ) -> List[StateChange]:
-    if len(claims) == 0:
-        return []
+    state_changes: List[StateChange] = []
 
-    state_changes: List[StateChange] = [
-        ContractReceiveNewTokenNetwork(
-            token_network_registry_address=token_network_registry_address,
-            token_network=TokenNetworkState(
-                address=token_network_address,
-                token_address=token_address,
-                network_graph=TokenNetworkGraphState(token_network_address),
-            ),
-            transaction_hash=TransactionHash(b""),
-            block_number=BlockNumber(0),
-            block_hash=BlockHash(b""),
-        )
-    ]
-
-    # TODO: check claim signature
     for claim in claims:
+
+        token_network_state = views.get_token_network_by_address(
+            chain_state=chain_state, token_network_address=claim.token_network_address
+        )
+        assert token_network_state is not None, "Token network does not exist"
+
         # If node is channel participant, create NettingChannelState
         if node_address == claim.owner or node_address == claim.partner:
             our_state = NettingChannelEndState(
@@ -164,7 +94,7 @@ def claims_to_blockchain_events(
                     token_network_address=claim.token_network_address,
                     channel_identifier=claim.channel_id(),
                 ),
-                token_address=token_address,
+                token_address=token_network_state.token_address,
                 token_network_registry_address=token_network_registry_address,
                 reveal_timeout=DEFAULT_REVEAL_TIMEOUT,
                 settle_timeout=DEFAULT_SETTLE_TIMEOUT,
@@ -184,12 +114,6 @@ def claims_to_blockchain_events(
                     block_hash=BlockHash(b""),
                 )
             )
-            transaction_channel_deposit = TransactionChannelDeposit(
-                participant_address=claim.owner,
-                contract_balance=claim.total_amount,
-                deposit_block_number=BlockNumber(1),
-            )
-
             state_changes.append(
                 ContractReceiveChannelDeposit(
                     canonical_identifier=CanonicalIdentifier(
@@ -197,7 +121,11 @@ def claims_to_blockchain_events(
                         token_network_address=claim.token_network_address,
                         channel_identifier=claim.channel_id(),
                     ),
-                    deposit_transaction=transaction_channel_deposit,
+                    deposit_transaction=TransactionChannelDeposit(
+                        participant_address=claim.owner,
+                        contract_balance=claim.total_amount,
+                        deposit_block_number=BlockNumber(1),
+                    ),
                     transaction_hash=TransactionHash(b""),
                     block_number=BlockNumber(1),
                     block_hash=BlockHash(b""),
@@ -224,15 +152,3 @@ def claims_to_blockchain_events(
             )
 
     return state_changes
-
-
-def get_state_changes_for_claims(
-    claims: List[Claim],
-    node_address: Address,
-    token_network_registry_address: TokenNetworkRegistryAddress,
-    token_network_address: TokenNetworkAddress,
-    token_address: TokenAddress,
-) -> List[StateChange]:
-    return claims_to_blockchain_events(
-        claims, node_address, token_network_registry_address, token_network_address, token_address
-    )
